+++ src/App.vue
@@ -1,0 +1,7 @@
+<template>
+  <SupplierLayout />
+</template>
+
+<script setup>
+import SupplierLayout from './layouts/SupplierLayout.vue'
+</script>